style(k6): remove redundant int64 conversions
ParseInt parses at bitSize 64 now, so the explicit int64() casts on its
results are unnecessary. Fixes golangci-lint unconvert failures.

+++ internal/resources/k6/data_source_k6_loadtest.go
@@ -111,7 +111,7 @@ func (d *loadTestDataSource) Read(ctx context.Context, req datasource.ReadReques
 		)
 		return
 	}
-	loadTestID := int64(intID)
+	loadTestID := intID
 
 	// Retrieve the load test attributes
 	ctx = context.WithValue(ctx, k6.ContextAccessToken, d.config.Token)
